Fix bafflingly stupid design decision
Previously, SAUVOLA_WINDOW_SIZE, a linear value, depended on the image
area, a quadratic value. I am baffled as to why I did this, and more
importantly, why didn't it break at least one of the 4 or 5 standard
image sizes the corpus consists of. Obviously the formula was
cherrypicked to work fine, but I cannot see how it worked so fine so
far.